+++ shunting-yard-replicator/src/test/java/com/hotels/shunting/yard/replicator/exec/event/aggregation/DefaultMetaStoreEventCompactorTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2018 Expedia Inc.
+ * Copyright (C) 2016-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,10 +21,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.powermock.api.mockito.PowerMockito.when;
 
-import static com.hotels.shunting.yard.common.event.EventType.ADD_PARTITION;
-import static com.hotels.shunting.yard.common.event.EventType.CREATE_TABLE;
-import static com.hotels.shunting.yard.common.event.EventType.DROP_PARTITION;
-import static com.hotels.shunting.yard.common.event.EventType.DROP_TABLE;
+import static com.expedia.apiary.extensions.receiver.common.event.EventType.ADD_PARTITION;
+import static com.expedia.apiary.extensions.receiver.common.event.EventType.CREATE_TABLE;
+import static com.expedia.apiary.extensions.receiver.common.event.EventType.DROP_PARTITION;
+import static com.expedia.apiary.extensions.receiver.common.event.EventType.DROP_TABLE;
 
 import java.util.List;
 
@@ -34,7 +34,8 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import com.hotels.shunting.yard.common.event.EventType;
+import com.expedia.apiary.extensions.receiver.common.event.EventType;
+
 import com.hotels.shunting.yard.replicator.exec.event.MetaStoreEvent;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -51,7 +52,7 @@ public class DefaultMetaStoreEventCompactorTest {
 
   @Test
   public void shouldReturnSameEventsIfEventsCannotBeMerged() {
-    List<MetaStoreEvent> events = asList(mockEvent(EventType.DROP_PARTITION), mockEvent(DROP_TABLE),
+    List<MetaStoreEvent> events = asList(mockEvent(DROP_PARTITION), mockEvent(DROP_TABLE),
         mockEvent(CREATE_TABLE));
     List<MetaStoreEvent> compactEvents = compactor.compact(events);
     assertThat(compactEvents).isEqualTo(events);