+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - TBD
+## [2.0.0] - TBD
 ### Added
 * Support for specifying target database and table name.
+
+### Changed
+* Extracted Apiary SQS receiver logic to library in [`apiary-receivers`](https://github.com/ExpediaInc/apiary-extensions/tree/master/apiary-receivers).
 
 ## [1.0.0] - 2019-03-19
 ### Added